FinancialAnalystAgentic/requirements.txt
# python packages needed
pandas
python-dotenv
streamlit
yfinance
rich
pyyaml
tabulate

# LLM APIs
google-genai

# agno & dependencies
agno
lancedb 
tantivy 
pypdf 
duckduckgo-search



